// src/app/models/customer.model.ts
export interface Customer {
  custId: number;
  companyName: string | null;
  contactName: string | null;
  contactTitle: string | null;
  address: string | null;
  city: string | null;
  region: string | null;
  postalCode: string | null;
  country: string | null;
  phone: string | null;
  fax: string | null;
}